ajustes

+++ core/models.py
@@ -37,9 +37,9 @@ class Empresa(models.Model):
     numero = models.CharField(max_length=10)
     cep = models.CharField(max_length=8)
 
-    latitude = models.DecimalField(max_digits=11, decimal_places=8, null=True, blank=True)
-    longitude = models.DecimalField(max_digits=11, decimal_places=8, null=True, blank=True)
-
+    latitude = models.FloatField(null=True, blank=True)
+    longitude = models.FloatField(null=True, blank=True)
+    
     telefone = models.CharField(max_length=20)
     email = models.EmailField()
     site = models.URLField(blank=True, null=True)
